backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==1.0.3
aiohttp==3.9.1
pdfplumber==0.10.3
python-docx==1.1.0
PyMuPDF==1.23.8
reportlab==4.0.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-magic==0.4.27
cryptography==41.0.7